requirements.txt
flask==3.0.0
langchain-community==0.0.13
langchain==0.1.0
chromadb==0.4.22
python-dotenv==1.0.0
requests==2.31.0
sentence-transformers==2.2.2
huggingface-hub==0.19.4 